requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
joblib==1.0.1
numpy==1.20.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.3
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
